+++ ideiasSol/src/components/employees/EmployeeList.jsx
@@ -1,9 +1,11 @@
 import React, { useState, useEffect } from "react";
 import "bootstrap/dist/css/bootstrap.min.css";
-import { getEmployee, updateEmployee, deleteEmployee, getGroups } from "../../services/api";
-import { getDepartments } from "../../services/api";
 import Select from "react-select";
+import { Link } from "react-router-dom";
 import CustomHeader from "../header/Header";
+import { getEmployee, updateEmployee, deleteEmployee, getGroups, getDepartments } from "../../services/api";
+import "bootstrap/dist/css/bootstrap.min.css";
+import { FaEdit, FaTrash, FaSave, FaTimes } from "react-icons/fa";
 
 const EmployeeList = () => {
   const [employees, setEmployees] = useState([]);
@@ -11,7 +13,6 @@ const EmployeeList = () => {
   const [editingEmployeeId, setEditingEmployeeId] = useState(null);
   const [editedEmployee, setEditedEmployee] = useState({
     name: "",
-    document: "",
     departmentId: "",
     groupId: "",
     active: true,
@@ -19,7 +20,7 @@ const EmployeeList = () => {
   });
   const [departmentList, setDepartmentList] = useState([]);
   const [selectedDepartment, setSelectedDepartment] = useState(null);
-  const [initialValue, setInitialValue] = useState("");
+  const [selectedGroup, setSelectedGroup] = useState(null);
 
   useEffect(() => {
     const fetchData = async () => {
@@ -38,23 +39,12 @@ const EmployeeList = () => {
     fetchData();
 
     (async () => {
-      const response = await getDepartments();
-      setDepartmentList(response.data);
+      const responseDepartments = await getDepartments();
+      setDepartmentList(responseDepartments.data);
+
+      const responseGroups = await getGroups();
+      setGroupsList(responseGroups.data);
     })();
-
-    (async () => {
-      const response = await getGroups();
-      setGroupsList(response.data);
-    })();
-  }, []);
-
-  useEffect(() => {
-    const fetchDepartments = async () => {
-      const response = await getDepartments();
-      setDepartmentList(response.data);
-    };
-
-    fetchDepartments();
   }, []);
 
   const fetchData = async () => {
@@ -78,18 +68,13 @@ const EmployeeList = () => {
     if (employeeToEdit) {
       setEditedEmployee({
         name: employeeToEdit.name,
-        document: employeeToEdit.document,
         departmentId: employeeToEdit.departmentId,
         groupId: employeeToEdit.groupId,
         groupName: employeeToEdit.groupName,
         depName: employeeToEdit.depName,
-
         active: employeeToEdit.active,
         type: employeeToEdit.type,
       });
-
-      setSelectedDepartment(options.find((option) => option.label === employeeToEdit.depName));
-      setInitialValue(employeeToEdit.groupId);
     }
   };
 
@@ -99,13 +84,14 @@ const EmployeeList = () => {
 
       if (response.status === 200) {
         setEditingEmployeeId(null);
-
         fetchData();
       } else {
         console.error("Erro ao editar funcionário");
+        alert(error.response.data.error);
       }
     } catch (error) {
       console.error("Erro ao editar funcionário", error);
+      alert(error.response.data.error);
     }
   };
 
@@ -115,11 +101,16 @@ const EmployeeList = () => {
 
   const handleDelete = async (employeeId) => {
     try {
+      const confirmDelete = window.confirm("Tem certeza que deseja excluir este funcionário?");
+
+      if (!confirmDelete) {
+        return;
+      }
+
       const response = await deleteEmployee(employeeId);
 
       if (response.status === 200) {
         alert("Funcionário excluído com sucesso!");
-
         fetchData();
       } else {
         console.error("Erro ao excluir funcionário");
@@ -130,16 +121,28 @@ const EmployeeList = () => {
     }
   };
 
-  const options = departmentList.map((department) => ({
+  const optionsDepartments = departmentList.map((department) => ({
     value: department.departmentId,
     label: department.name,
   }));
 
+  const optionsGroups = groupsList.map((group) => ({
+    value: group.groupId,
+    label: group.name,
+  }));
+
   return (
     <div>
-      <CustomHeader></CustomHeader>
+      <CustomHeader />
       <div className="container mt-4">
-        <h1 className="text-center">Lista de Funcionários</h1>
+        <div className="row text-center justify-content-end">
+          <h1>Lista de Funcionários</h1>
+          <div className="col-auto justify-content-end">
+            <Link to="/insertEmployee" className="btn btn-primary">
+              +
+            </Link>
+          </div>
+        </div>
         <table className="table">
           <thead>
             <tr>
@@ -150,7 +153,7 @@ const EmployeeList = () => {
               <th>Grupo</th>
               <th>Status</th>
               <th>Tipo</th>
-              <th>Ações</th>
+              <th></th>
             </tr>
           </thead>
           <tbody>
@@ -169,31 +172,20 @@ const EmployeeList = () => {
                     employee.name
                   )}
                 </td>
-                <td>
-                  {editingEmployeeId === employee.employeeId ? (
-                    <input
-                      type="text"
-                      value={editedEmployee.document}
-                      onChange={(e) => setEditedEmployee({ ...editedEmployee, document: e.target.value })}
-                      className="form-control"
-                    />
-                  ) : (
-                    employee.document
-                  )}
-                </td>
+                <td>{employee.document}</td>
                 <td>
                   {editingEmployeeId === employee.employeeId ? (
                     <Select
                       name="departmentId"
                       id="departmentId"
-                      value={editedEmployee.departmentId}
-                      defaultValue={selectedDepartment}
+                      value={optionsDepartments.find((option) => option.value === editedEmployee.departmentId)}
                       onChange={(selectedOption) => {
                         setSelectedDepartment(selectedOption);
                         setEditedEmployee({ ...editedEmployee, departmentId: selectedOption.value });
                       }}
-                      options={options}
+                      options={optionsDepartments}
                       isSearchable
+                      required
                     />
                   ) : (
                     employee.depName
@@ -201,25 +193,36 @@ const EmployeeList = () => {
                 </td>
                 <td>
                   {editingEmployeeId === employee.employeeId ? (
-                    <select
-                      onChange={(e) => setEditedEmployee({ ...editedEmployee, groupId: e.target.value })}
+                    <Select
                       name="groupId"
                       id="groupId"
-                      value={editedEmployee.groupId}
-                      className="form-control"
-                    >
-                      <option value="">Selecione um grupo</option>
-                      {groupsList.map((groups) => (
-                        <option key={groups.groupId} value={groups.groupId}>
-                          {groups.name}
-                        </option>
-                      ))}
-                    </select>
+                      value={optionsGroups.find((option) => option.value === editedEmployee.groupId)}
+                      onChange={(selectedOption) => {
+                        setSelectedGroup(selectedOption);
+                        setEditedEmployee({ ...editedEmployee, groupId: selectedOption.value });
+                      }}
+                      options={optionsGroups}
+                      isSearchable
+                      required
+                    />
                   ) : (
                     employee.groupName
                   )}
                 </td>
-                <td>{employee.active ? "Ativo" : "Inativo"}</td>
+                <td>
+                  {editingEmployeeId === employee.employeeId ? (
+                    <select
+                      onChange={(e) => setEditedEmployee({ ...editedEmployee, active: e.target.value })}
+                      value={editedEmployee.active}
+                      className="form-control"
+                    >
+                      <option value="1">Ativo</option>
+                      <option value="0">Inativo</option>
+                    </select>
+                  ) : (
+                    <td>{employee.active ? "Ativo" : "Inativo"}</td>
+                  )}
+                </td>
                 <td>
                   {editingEmployeeId === employee.employeeId ? (
                     <select
@@ -238,19 +241,21 @@ const EmployeeList = () => {
                   {editingEmployeeId === employee.employeeId ? (
                     <div>
                       <button className="btn btn-success btn-sm mr-2" onClick={handleSave}>
-                        Salvar
+                        <FaSave />
                       </button>
+                      &nbsp;
                       <button className="btn btn-danger btn-sm" onClick={handleCancelEdit}>
-                        Cancelar
+                        <FaTimes />
                       </button>
                     </div>
                   ) : (
                     <div>
                       <button className="btn btn-primary btn-sm mr-2" onClick={() => handleEdit(employee.employeeId)}>
-                        Editar
+                        <FaEdit />
                       </button>
+                      &nbsp;
                       <button className="btn btn-danger btn-sm" onClick={() => handleDelete(employee.employeeId)}>
-                        Excluir
+                        <FaTrash />
                       </button>
                     </div>
                   )}
